Reorganise classes into packages
