+++ src/main/java/bob/Storage.java
@@ -1,5 +1,7 @@
 package bob;
 
+import utilities.Printer;
+
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
