+++ src/main/java/commands/Command.java
@@ -1,4 +1,9 @@
-package bob;
+package commands;
+
+import bob.ExecutionStack;
+import bob.Storage;
+import bob.TaskList;
+import bob.Ui;
 
 /**
  * Abstract class representing a command that can be executed.
@@ -13,9 +18,9 @@ public abstract class Command {
      * @param ui the {@code Ui} to interact with the user.
      * @param storage the {@code Storage} to save or load tasks.
      */
-    protected abstract String execute(TaskList tasks, Ui ui, Storage storage, ExecutionStack execStack);
+    public abstract String execute(TaskList tasks, Ui ui, Storage storage, ExecutionStack execStack);
 
-    protected abstract String undo(TaskList tasks, Ui ui, Storage storage);
+    public abstract String undo(TaskList tasks, Ui ui, Storage storage);
 
     /**
      * Determines if the command is an exit command.
@@ -24,7 +29,7 @@ public abstract class Command {
      *
      * @return {@code true} if the command is an exit command; {@code false} otherwise.
      */
-    protected boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
