+++ src/main/java/commands/FindDateCommand.java
@@ -1,8 +1,11 @@
-package bob;
+package commands;
+
+import bob.*;
+import utilities.Printer;
 
 public class FindDateCommand extends Command {
-    protected static String[] params = new String[] { "findDate" };
-    protected static int paramCount = 1;
+    public static String[] params = new String[] { "findDate" };
+    public static int paramCount = 1;
     private final String date;
 
     public FindDateCommand(String date) {
