+++ src/main/java/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package bob;
+package gui;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
