+++ src/main/java/gui/Main.java
@@ -1,5 +1,6 @@
-package bob;
+package gui;
 
+import bob.Bob;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
