+++ src/main/java/gui/MainWindow.java
@@ -1,5 +1,7 @@
-package bob;
+package gui;
 
+import bob.Bob;
+import bob.Ui;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -7,10 +9,6 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import javafx.application.Platform;
-
-import java.util.concurrent.TimeUnit;
-import java.lang.InterruptedException;
 
 
 /**
