+++ src/main/java/tasks/Event.java
@@ -1,4 +1,4 @@
-package bob;
+package tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
